add delete member feature

+++ src/main/java/com/library/LibMgmt2025/controller/MemberController.java
@@ -2,6 +2,7 @@ package com.library.LibMgmt2025.controller;
 
 import com.library.LibMgmt2025.dto.MemberDto;
 import com.library.LibMgmt2025.dto.StaffDto;
+import com.library.LibMgmt2025.exception.MemberNotFoundException;
 import com.library.LibMgmt2025.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -21,13 +22,25 @@ public class MemberController {
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> addMember(@RequestBody MemberDto memberDto){
+        if (memberDto == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         memberService.saveMember(memberDto);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
     @DeleteMapping
     public ResponseEntity<Void> deleteMember(@RequestParam("memberId") String memberId){
-        memberService.deleteMember(memberId);
-        return ResponseEntity.noContent().build();
+        if (memberId == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try {
+            memberService.deleteMember(memberId);
+            return ResponseEntity.noContent().build();
+        }catch (MemberNotFoundException e) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PatchMapping(value = "/{memberId}", consumes = MediaType.APPLICATION_JSON_VALUE)
